robot, bullets and explosions are level bound

+++ src/bullet.rs
@@ -4,6 +4,7 @@ use bevy::{
     math::{vec2, vec3},
     prelude::*,
 };
+use bevy_ecs_ldtk::LevelIid;
 use bevy_rapier2d::prelude::*;
 
 use crate::{
@@ -41,27 +42,34 @@ fn spawn_bullet(
     mut cmd: Commands,
     assets: Res<TextureAssets>,
     mut ev_fire: EventReader<FireEvent>,
+    q_level: Query<Entity, With<LevelIid>>,
 ) {
     for ev in ev_fire.read() {
-        let angle = vec2(1.0, 0.0).angle_between(ev.to_target.truncate() - ev.from_pos.truncate());
-        let quat = Quat::from_rotation_z(angle);
-        cmd.spawn((
-            Bullet {
-                timer: Timer::new(BULLET_LIFETIME, TimerMode::Once),
-                direction: quat.mul_vec3(vec3(1.0, 0.0, 0.0)).truncate(),
-                angle,
-            },
-            SpriteBundle {
-                texture: assets.bullet.clone(),
-                transform: Transform::from_scale(Vec3::splat(0.5))
-                    .with_rotation(quat)
-                    .with_translation(vec3(ev.from_pos.x, ev.from_pos.y, 3.)),
-                ..Default::default()
-            },
-            // Collider::capsule_x(6.0, 2.0),
-            // Sensor,
-            ev.team,
-        ));
+        if let Ok(e_level) = q_level.get_single() {
+            let angle =
+                vec2(1.0, 0.0).angle_between(ev.to_target.truncate() - ev.from_pos.truncate());
+            let quat = Quat::from_rotation_z(angle);
+            let e_bullet = cmd
+                .spawn((
+                    Bullet {
+                        timer: Timer::new(BULLET_LIFETIME, TimerMode::Once),
+                        direction: quat.mul_vec3(vec3(1.0, 0.0, 0.0)).truncate(),
+                        angle,
+                    },
+                    SpriteBundle {
+                        texture: assets.bullet.clone(),
+                        transform: Transform::from_scale(Vec3::splat(0.5))
+                            .with_rotation(quat)
+                            .with_translation(vec3(ev.from_pos.x, ev.from_pos.y, 3.)),
+                        ..Default::default()
+                    },
+                    // Collider::capsule_x(6.0, 2.0),
+                    // Sensor,
+                    ev.team,
+                ))
+                .id();
+            cmd.entity(e_level).add_child(e_bullet);
+        }
     }
 }
 
@@ -83,12 +91,12 @@ fn move_bullet(
 fn check_bullet_hit(
     mut cmd: Commands,
     rapier_context: Res<RapierContext>,
-    q_bullet: Query<(Entity, &Bullet, &GlobalTransform, &Team)>,
+    q_bullet: Query<(Entity, &Bullet, &GlobalTransform, &Transform, &Team)>,
     mut q_other: Query<(&mut Life, &Team)>,
     mut ev_explosion: EventWriter<ExplosionEvent>,
 ) {
     let shape = Collider::capsule_x(6.0, 2.0);
-    for (e_bullet, bullet, bullet_gtr, bullet_team) in &q_bullet {
+    for (e_bullet, bullet, bullet_gtr, bullet_tr, bullet_team) in &q_bullet {
         let filter = QueryFilter {
             groups: Some(match *bullet_team {
                 Team::Player => coll_groups(
@@ -114,7 +122,7 @@ fn check_bullet_hit(
                         cmd.entity(e_bullet).despawn_recursive();
                         life.curr_hp -= 1.0;
                         ev_explosion.send(ExplosionEvent {
-                            location: bullet_gtr.translation().truncate(),
+                            location: bullet_tr.translation.truncate(),
                             particle_radius: 2.,
                             spread: 3.,
                             particle_speed: bullet.direction / 2.,
